fix: move saveToFile to background for Firefox support
Firefox does not support `saveAs` in a popup, so `saveToFile` has been moved
to the background script. A message listener is added to handle save requests.
